[topic-54] Perform initial component refactor
