+++ package.json
@@ -20,10 +20,12 @@
     "test": "jest --passWithNoTests"
   },
   "dependencies": {
+    "@expo/webpack-config": "^19.0.0",
     "@react-native-community/datetimepicker": "7.2.0",
     "@react-native-picker/picker": "2.4.10",
     "@react-navigation/material-top-tabs": "^6.6.4",
     "@react-navigation/native": "^6.1.8",
+    "@types/react": "~18.2.14",
     "buffer": "^6.0.3",
     "expo": "^49.0.13",
     "expo-apple-authentication": "~6.1.0",
@@ -34,14 +36,18 @@
     "jest-expo": "^49.0.0",
     "lodash": "^4.17.21",
     "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-native": "0.72.5",
     "react-native-pager-view": "6.2.0",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "react-native-tab-view": "^3.5.2"
+    "react-native-tab-view": "^3.5.2",
+    "react-native-web": "~0.19.6",
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
+    "@types/lodash": "^4.14.202",
     "husky": "^8.0.0",
     "lint-staged": "^14.0.1",
     "prettier": "^3.0.3",
